montecarlo working and the result is pi

+++ Student_Cuda/src/cpp/core/04_Montecarlo/01_object/host/Montecarlo.h
@@ -27,7 +27,7 @@ class Montecarlo{
 
 	int nbFlechette;
 
-	int M;
+	float M;
 	float a, b;
 
 	int nbFlechetteDessous;
